requirements.txt
python-docx
faiss-cpu
ftfy
streamlit 
sentence-transformers
